backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2
google-genai>=1.0.0
groq>=0.12.0
pydantic>=2.0.0
